refactor(all): moving classes to models and controllers packages
