+++ src/main/java/com/cs3365/rpgdapi/controllers/WeaponsService.java
@@ -1,4 +1,7 @@
-package com.cs3365.rpgdapi.weapons;
+package com.cs3365.rpgdapi.controllers;
+
+import com.cs3365.rpgdapi.models.Weapon;
+import com.cs3365.rpgdapi.models.WeaponEntity;
 
 import java.util.UUID;
