+++ src/main/java/com/cs3365/rpgdapi/models/Weapon.java
@@ -1,4 +1,6 @@
-package com.cs3365.rpgdapi.weapons;
+package com.cs3365.rpgdapi.models;
+
+import com.cs3365.rpgdapi.controllers.WeaponsException;
 
 public class Weapon {
